+++ packages/core/src/components/input-range/input-range.interfaces.ts
@@ -1,0 +1,7 @@
+export interface InputRangeChangeEvent {
+  originalEvent: Event;
+  value: number | undefined;
+  min: number;
+  max: number;
+  step: number;
+}