+++ packages/dso-toolkit/src/components/input-range/index.ts
@@ -1,0 +1,3 @@
+export * from "./input-range.args.js";
+export * from "./input-range.models.js";
+export * from "./input-range.stories-of.js";