+++ packages/dso-toolkit/src/components/input-range/input-range.models.ts
@@ -1,0 +1,18 @@
+export interface InputRange {
+  min?: number;
+  max?: number;
+  value?: number;
+  step?: number;
+  label?: string;
+  unit: string;
+  description?: string;
+  dsoChange: (e: InputRangeChangeEvent) => void;
+}
+
+export interface InputRangeChangeEvent {
+  originalEvent: Event;
+  value: number | undefined;
+  min: number;
+  max: number;
+  step: number;
+}